Move some of the gui elements into a separate package
