+++ src/vampireEditor/print/Backgrounds.java
@@ -25,7 +25,7 @@ package vampireEditor.print;
 import vampireEditor.VampireEditor;
 import vampireEditor.entity.Character;
 import vampireEditor.entity.EntityException;
-import vampireEditor.gui.Font;
+import vampireEditor.gui.utility.Font;
 import vampireEditor.print.utility.StringProperties;
 
 /**
